TEST: Migrate WooCommerceProvider tests from Feature to Unit directory and enhance test coverage
